### 2019-06-13 #1502 @arfon
> Failed to discover a valid open source license.

Not sure why @whedon isn't happy - GitHub seems to think this is MIT-licensed (https://github.com/for-ai/rl/blob/master/LICENSE).
