### 2019-06-18 #1502 @mbobra
@alexanderimanicowenrivers I'm sorry for the delay -- I'm still working on finding reviewers for this submission. I don't know the average review time for a JOSS submission, but it can be short (less than one week) once the handling editor finds reviewers. I've messaged a ton of potential reviewers and it looks like a busy time -- but I'll get there. Thank you!
