### 2019-06-20 #1502 @desilinguist
I am familiar with TF but not very familiar with reinforcement learning. If
you don’t find anyone else better suited, I’ll take a look.

On Thu, Jun 20, 2019 at 5:56 AM Chris Markiewicz <notifications@github.com>
wrote:

> Sorry, no TF experience.
>
