### 2019-06-20 #1502 @rougier
I'm a bit busy right now (and I'm just the opposite of @desilinguist, familiar with RL but no TL).
One of our PhD student here (@cervere) might be able to review if needed.
