### 2019-06-20 #1502 @mbobra
@desilinguist There is no need to be an expert -- according to the JOSS docs, a reviewer only needs "[sufficient expertise](https://joss.readthedocs.io/en/latest/editing.html#finding-reviewers)" in the field of submission. The [reviewer checklist](https://joss.readthedocs.io/en/latest/review_checklist.html) covers general software, documentation, and functionality (i.e. confirming the functional claims of the software). I think you'd be more than qualified to do this job. Do you mind reviewing this submission?

@cervere Do you have time to review this submission?
