### 2019-06-20 #1502 @cervere
@mbobra Sorry, cannot make time at this moment.
(Had a quick run through, so if I am allowed make some unsolicited remarks :
I would have liked to spend some time on certain aspects - particularly because the work claims 'generic' - for e.g, envs outside OpenAI? more complex agents with multiple objectives? etc..
And as observed in other submissions, I hope the authors have addressed the important question well : What is the claim? 'Generic' nature from RL pov or from a utility for tensorflow pov..)
Cheers
