### 2019-06-21 #1502 @alexanderimanicowenrivers
> Sorry, no TF experience.

Hey @effigies , I don't think TF experience is necessary, the most important aspects are the way its structured and the algorithms we provide, thus if you know RL that should me more to suffice a review :)
