### 2019-06-21 #1502 @alexanderimanicowenrivers
> @mbobra Sorry, cannot make time at this moment.
> (Had a quick run through, so if I am allowed make some unsolicited remarks :
> I would have liked to spend some time on certain aspects - particularly because the work claims 'generic' - for e.g, envs outside OpenAI? more complex agents with multiple objectives? etc..
> And as observed in other submissions, I hope the authors have addressed the important question well : What is the claim? 'Generic' nature from RL pov or from a utility for tensorflow pov..)
> Cheers

So generic in both terms, 1 we provide integration to external gyms (CoinRun, Atari and OpenAI GYM) and its generic in terms of a RL point of view, abstracting out the relevant methods allowing re-use across both model-free and model-based algorithms.

It was also generic in the sense that when we could, we wrote non differentiable functions in numpy (e.g memory), which allowed me recently to port the whole codebase over to PyTorch with minimal effort (just changing the models, train.py and algos folder).
