### 2019-06-21 #1502 @jkahn
currently without time to contribute here. Wish I could; I'm excited to see
the sort of work described by the title.

On Wed, Jun 19, 2019 at 9:50 PM Alexander Cowen-Rivers <
notifications@github.com> wrote:

> @desilinguist <https://github.com/desilinguist> , @rougier
> <https://github.com/rougier>, @sealhuang <https://github.com/sealhuang>,
> @ahurriyetoglu <https://github.com/ahurriyetoglu>, @betatim
> <https://github.com/betatim>, @jkahn <https://github.com/jkahn>, @nirum
> <https://github.com/nirum>, @arokem <https://github.com/arokem>
>
