### 2019-06-22 #1502 @alexanderimanicowenrivers
@mbobra can we assign the review to. @desilinguist  please?

https://github.com/openjournals/joss-reviews/issues/1502#issuecomment-503985440
