### 2019-06-24 #1524 @whedon
[REVIEW]: Generic reinforcement learning codebase in TensorFlow
**Submitting author:** @alexanderimanicowenrivers (<a href="http://orcid.org/0000-0002-2669-9513">Alexander I. Cowen-Rivers</a>)
**Repository:** <a href="https://github.com/for-ai/rl" target ="_blank">https://github.com/for-ai/rl</a>
**Version:** 1.0.0
