### 2019-06-24 #1524 @alexanderimanicowenrivers
Thanks for agreeing to review, and I have already addressed these comments in my reply here https://github.com/openjournals/joss-reviews/issues/1502#issuecomment-504362129
