### 2019-07-02 #1524 @alexanderimanicowenrivers
Let me know if there is anything you would like clarified! @desilinguist @paragkulkarni11
