### 2019-07-12 #1524 @paragkulkarni11
Hi,
I am sorry for late review. In my opinion, documentation could be more elaborate adhering to guidelines given above. In particular , I will expect to be more specific when we say that past work was more general in nature. This will clearly specify and support your claim. Also can we see any real life example of this codebase implementation ? Kindly include it in your documentation. In case,, if possible please add most frequently used words of your paper. I am interested to see references of the paper used for this research. Please let me know if any point is not clear from above discussion. Thanks.
