### 2019-07-14 for-ai/rl#8 @desilinguist
Improve installation
The current way of installation by running `./setup.sh` has a couple of issues:

1. It assumes that a macOS user already has homebrew installed. This is not always true - for example, I use [macports](macports.org) and not homebrew This means that the installation fails for me.

2. It also installs all the python packages in the main python environment. This is not ideal since it might cause conflicts with package versions that might be there for a reason.

I would recommend that you revamp your installation instructions to explicitly list out the various setups that are in `./setup.sh` to make it easier for users to see what they are doing, to allow using macports instead of homebrew, and to include a step to create a virtual environment before installing the various python packages.

If you can do the latter using `conda`, that would be even better since the `conda-forge` channel there might even include some of the non-python dependencies (`ffmpeg`, `pkg-config`) which means the user does not need to worry about causing conflicts there either and those can be contained in the virtual environment as well.
