### 2019-07-14 for-ai/rl#9 @desilinguist
Improve documentation to include novice users
There's no documentation for this project at all which is less than ideal. Basically, this project assumes that you are extremely well-versed in reinforcement learning and can figure out how to run this codebase yourself by looking at the code. For example, I am a complete novice and I couldn't figure out how to easily try something out using this codebase once I had finished installation. Even the quick start section in the README is not a complete example.

I would recommend that you create actual documentation on readthedocs.org that explains what RL is, what this codebase does for RL compared to other things, and provide a complete end-to-end example on how to use this codebase. The rest of the API and the modules should also be documented clearly.
