### 2019-07-14 #1524 @desilinguist
I have completed my review and I think while the goal of the project is admirable – providing a generic codebase for reinforcement learning in Tensorflow that also provides nice logging and visualization features – the setup process and the overall documentation both leave a lot to be desired. I have filed [two](https://github.com/for-ai/rl/issues/8) [issues](https://github.com/for-ai/rl/issues/9) with suggestions on how to address both of these.

The software paper does have a bit of motivation but it isn't really clear what the target demographic is - is it RL researchers who are already experts and this is a way for them to run experiments or are novice users also included? I got the impression from my interaction with the authors that it was the latter but I was quite lost.

There also doesn't seem to be any real tests - the Travis CI script just runs a bunch of python `train.py` commands with a bunch of parameters to see whether anything would fail? Am I misreading that?
