### 2019-07-15 #1524 @mbobra
👋 @alexanderimanicowenrivers Can you please address the comments by @paragkulkarni11 and @desilinguist? Please let me know if you want any help with this.
