### 2019-07-20 0518ee9
Update paper.md (#10)
Nice work!
