### 2019-07-20 for-ai/rl#8 @alexanderimanicowenrivers
I believe this is completed now. Thanks!
