### 2019-07-20 for-ai/rl#9 @alexanderimanicowenrivers
This is also now complete https://github.com/for-ai/rl/pull/10

Thanks!
