### 2019-07-20 #1524 @alexanderimanicowenrivers
Thanks for the reviews @desilinguist and @paragkulkarni11

@desilinguist - We have resolved and closed both issues, providing clearer documentation with working examples, as well as an improved initialisation script. We are indeed targetting the RL researcher, rather than a novice. The idea is for this codebase to be used specifically for RL research, rather than for industrial systems (although it could be used for this too)!.

@paragkulkarni11  - Thanks for the feedback, as mentioned above we have addressed the issues of working example and clearer documentation. We have yet to have our codebase used in research (we only released it a few months ago) --- but are currently collaborating using the codebase with various people.

The reason our tests are just run scripts is due to the high stochasticity of experiments. However, we are working on developing deterministic settings which we can rigorously test. We are awaiting adding in a few more tests on the non-stochastic components.

I believe we have addressed all major issues with the review, is there anything else we can help you with?

Thanks!

Alex and the For.ai team :)
