### 2019-07-20 dc2e005
Update paper.md
