### 2019-07-24 #1524 @mbobra
@desilinguist @paragkulkarni11 Do you recommend this paper for publication after @alexanderimanicowenrivers' improvements or are there still some outstanding issues? I see that the checklists are not complete and I would like some guidance on how you would like to proceed from here.
