### 2019-07-24 #1524 @desilinguist
Thanks @alexanderimanicowenrivers for the modifications! Unfortunately, I am traveling to a conference outside the country today and won't be back until August 5th. I will be happy to re-review once I am back.
