### 2019-08-01 5a1ce0e
adding tests and benchmark
