### 2019-08-01 for-ai/rl#9 @alexanderimanicowenrivers
https://rl-codebase.readthedocs.io/en/latest/
