### 2019-08-01 c7ed9bb
Update README.md
Adding in extended installation instructions.
