+++ README.md
@@ -26,10 +26,23 @@ Example for recorded envrionment on various RL agents.
 ![MountainCar-v0](gif/mountaincar.gif)|![Pendulum-v0](gif/pendulum.gif)|![VideoPinball-v0](gif/pinball.gif)|![Tennis-v0](gif/tennis.gif)
 
 ### Requirements
-- Run setup scripts to install all dependencies and environments
+It is recommended to install the codebase in a virtual environment ([virtualenv](https://pypi.org/project/virtualenv/) or [conda](https://conda.io/en/latest/)). 
+
+### Quick install
+Configure `use_gpu` and (if on OSX) `mac_package_manager` (either [macports](https://www.macports.org) or [homebrew](https://brew.sh)) params in `setup.sh`, then run it as
+```bash
+sh setup.sh
 ```
-./setup.sh
-```
+
+### Manual setup
+You need to install the following for your system:
+
+- [TensorFlow](https://www.tensorflow.org/install)
+- [OpenAI Gym](https://gym.openai.com/docs/#installation)
+    - Atari `pip install 'gym[atari]'`
+- [OpenAI CoinRun](https://github.com/openai/coinrun)
+- [FFmpeg](https://ffmpeg.org/download.html)
+- Additional python packages `pip install -r ../requirements.txt`
 
 ### Quick Start
 ```
