### 2019-08-01 #1524 @alexanderimanicowenrivers
Dear reviewer,

We thank you for your comments and suggestions and believe we have sufficiently addressed all points.

Regarding your comment on the installation script, we have added to the file `setup.sh` such that the installation script now supports Homebrews on macOS. In addition, we have written additional logging lines that will allow the user to know exactly what is being installed and any errors if they occur. Furthermore, we have added to the file `README.md` additional information on how the user can manually install all dependencies should they choose to do so.

Regarding your comment on additional documentation, we have added to the files `README.md` and `paper.md`. Furthermore, we have created a readthedocs documentation paper, located at: https://rl-codebase.readthedocs.io/en/latest/. Here, we detail the different modules provided as well as a tutorial and basic usage. In the tutorial, we have outlined the use of a Conda environment to run experiments, as you requested us to do.

Regarding your comment on tests, we have under the directory `\tests` provided testing for memory and models.

Thanks for the great feedback @desilinguist and @mbobra!

Sincerely,
the FOR.ai team
