### 2019-08-02 e0dabb3
fixes and edits
