### 2019-08-04 91c7572
Update paper.md
Adding Sheldon into the JOSS submission.
