### 2019-08-04 #1524 @alexanderimanicowenrivers
I hope you had a nice holiday @desilinguist , I look forward to hearing back from you and @paragkulkarni11
