+++ docs/source/references.bib
@@ -1,0 +1,139 @@
+@article{schulman2017proximal,
+  title={Proximal policy optimization algorithms},
+  author={Schulman, John and Wolski, Filip and Dhariwal, Prafulla and Radford, Alec and Klimov, Oleg},
+  journal={arXiv preprint arXiv:1707.06347},
+  year={2017}
+}
+
+@misc{sutton:1998,
+  title={Introduction to reinforcement learning. Vol. 135},
+  author={Sutton, Richard S and Barto, Andrew G},
+  year={1998},
+  publisher={Cambridge: MIT press}
+}
+
+@article{dhariwal:2017,
+  title={Openai baselines},
+  author={Dhariwal, Prafulla and Hesse, Christopher and Klimov, Oleg and Nichol, Alex and Plappert, Matthias and Radford, Alec and Schulman, John and Sidor, Szymon and Wu, Yuhuai and Zhokhov, Peter},
+  journal={GitHub, GitHub repository},
+  year={2017}
+}
+
+@article{schaarschmidt:2017,
+  title={Tensorforce: A tensorflow library for applied reinforcement learning},
+  author={Schaarschmidt, Michael and Kuhnle, Alexander and Fricke, Kai},
+  journal={Web page},
+  year={2017}
+}
+
+@article{liang:2017,
+  title={Ray rllib: A composable and scalable reinforcement learning library},
+  author={Liang, Eric and Liaw, Richard and Nishihara, Robert and Moritz, Philipp and Fox, Roy and Gonzalez, Joseph and Goldberg, Ken and Stoica, Ion},
+  journal={arXiv preprint arXiv:1712.09381},
+  year={2017}
+}
+
+@article{caspi:2017,
+  title={Reinforcement Learning Coach.(Dec. 2017)},
+  author={Caspi, Itai and Leibovich, Gal and Novik, Gal},
+  journal={URL https://doi. org/10.5281/zenodo},
+  volume={1134899},
+  year={2017}
+}
+
+@inproceedings{abadi:2016,
+  title={Tensorflow: A system for large-scale machine learning},
+  author={Abadi, Mart{\'\i}n and Barham, Paul and Chen, Jianmin and Chen, Zhifeng and Davis, Andy and Dean, Jeffrey and Devin, Matthieu and Ghemawat, Sanjay and Irving, Geoffrey and Isard, Michael and others},
+  booktitle={12th $\{$USENIX$\}$ Symposium on Operating Systems Design and Implementation ($\{$OSDI$\}$ 16)},
+  pages={265--283},
+  year={2016}
+}
+
+@article{brockman:2016,
+  title={Openai gym},
+  author={Brockman, Greg and Cheung, Vicki and Pettersson, Ludwig and Schneider, Jonas and Schulman, John and Tang, Jie and Zaremba, Wojciech},
+  journal={arXiv preprint arXiv:1606.01540},
+  year={2016}
+}
+
+@misc{kerasrl:2019, title={keras-rl/keras-rl}, url={https://github.com/keras-rl/keras-rl}, journal={GitHub}, author={Keras-Rl, Matthias Plappert}, year={2019}, month={Mar}}
+
+@article{castro:2018,
+  title={Dopamine: A research framework for deep reinforcement learning},
+  author={Castro, Pablo Samuel and Moitra, Subhodeep and Gelada, Carles and Kumar, Saurabh and Bellemare, Marc G},
+  journal={arXiv preprint arXiv:1812.06110},
+  year={2018}
+}
+
+@article{mnih:2013,
+  title={Playing atari with deep reinforcement learning},
[…]
