+++ docs/source/installation.rst
@@ -1,0 +1,9 @@
+============
+Installation
+============
+
+Install all dependencies by running::
+
+    $ ./setup.sh
+
+This installation script currently supports Linux and macOS. For macOS users, this installation script supports both Homebrew and Macports. In the event the script fails, we recommend installing all dependencies within a Conda environment.
