### 2019-08-05 78216a1
Adding pointers to the readthedoc documentation.
