### 2019-08-05 #1524 @desilinguist
@alexanderimanicowenrivers thanks so much for considering my suggestions! I think there has been great progress. However, I still see a couple of issues:

1. I don't see the changes you described in the README in `setup.sh`. For example, I don't see the `mac_package_manaager` variable anywhere in `setup.sh` in the `master` branch?

2.  I don't see the external documentation linked anywhere from the repository README? It'd be nice to add a readthedocs badge from shields.io right at the top of the README next to the passing builds.

3. Thanks for adding more explicit tests. However, I see lots of deprecation warnings in the [test output](https://travis-ci.org/for-ai/rl). Perhaps you can address those in the codebase or if you don't think they apply, you can turn off the deprecation warnings?
