### 2019-08-05 8e81a85
adding the documentation badge
