### 2019-08-05 6af9896
Adding the source code of documentation (#17)
* transferring the source code of documentation

* Adding pointers to the readthedoc documentation.

* adding the documentation badge
