+++ docs/requirements.txt
@@ -1,0 +1,3 @@
+sphinx==2.1.2
+sphinx_rtd_theme
+sphinxcontrib-bibtex
