+++ docs/source/index.rst
@@ -1,0 +1,13 @@
+
+Generic Reinforcment Learning Codebase
+=======================================
+
+
+Contents:
+
+.. toctree::
+   :maxdepth: 4
+   :glob:
+
+   *
+
