+++ docs/source/contents.rst
@@ -1,0 +1,110 @@
+Overview
+=========
+
+Abstract
+--------
+
+Vast reinforcement learning (RL) research groups, such as DeepMind and OpenAI, have their internal (private) reinforcement learning codebases, which enable quick prototyping and comparing of ideas to many SOTA methods. We argue the five fundamental properties of a sophisticated research codebase are; modularity, reproducibility, many RL algorithms pre-implemented, speed and ease of running on different hardware/ integration with visualization packages. Currently, there does not exist any RL codebase, to the author's knowledge, which contains all the five properties, particularly with TensorBoard logging and abstracting away cloud hardware such as TPU's from the user. The codebase aims to help distil the best research practices into the community as well as ease the entry access and accelerate the pace of the field.
+
+Related Work
+------------
+
+There are currently various implementations available for reinforcement learning codebase like OpenAI baselines :cite:`dhariwal:2017`, Stable baselines :cite:`hill:2019`, Tensorforce :cite:`schaarschmidt:2017`, Ray rllib :cite:`liang:2017`, Intel Coach :cite:`caspi:2017`, Keras-RL :cite:`kerasrl:2019`, Dopamine baselines :cite:`castro:2018` and TF-Agents :cite:`guadarramatf`. Ray rllib :cite:`liang:2017` is amongst the strongest of existing RL frameworks, supporting; distributed operations, TensorFlow :cite:`abadi:2016`, PyTorch :cite:`paszke:2017` and multi-agent reinforcement learning (MARL). Unlike Ray rllib, we choose to focus on Tensorflow support, allowing us to integrate specific framework visualisation and experiment tracking into our codebase. On top of this, we are developing a Kuberenetes script for MacOS and Linux users to connect to any cloud computing platform, such as Google TPU’s, Amazon AWS etc. Most other frameworks are plagued with problems like usability issues (difficult to get started and increment over), very little modularity in code (no/ little hierarchy and code reuse), no asynchronous training support, weak support for TensorBoard logging and so on. All these problems are solved by our project, which is a generic codebase built for reinforcement learning (RL) research in Tensorflow :cite:`schaarschmidt:2017`, with favoured RL agents pre-implemented as well as integration with OpenAI Gym :cite:`brockman:2016` environment focusing on quick prototyping and visualisation.
+
+Deep Reinforcement Learning Reinforcement learning refers to a paradigm in artificial intelligence where an agent performs a sequence of actions in an environment to maximise rewards :cite:`sutton:1998`. It is in many ways more general and challenging than supervised learning since it requires no labels to train on; instead, the agent interacts continuously with the environment, gathering more and more data and guiding its learning process.
+
+Introduction: for-ai/rl
+-----------------------
+
+Further to the core ideas mentioned in the beginning, a good research codebase should enable good development practices such as continually checkpointing the model's parameters as well as instantly restoring them to the latest checkpoint when available. Moreover, it should be composed of simple, interchangeable building blocks, making it easy to understand and to prototype new research ideas.
+
+We will first introduce the framework for this project, and then we will detail significant components. Lastly, we will discuss how to get started with training an agent under this framework.
+
+This codebase allows training RL agents by a training script as simple as the below for loop. ::
+
+	$ for epoch in range(epochs):
+    	$ state = env.reset()
+		$ for step in range(max_episode_steps):
+        	$ last_state = state
+        	$ action = agent.act(state)
+        	$ state, reward, done = env.step(action)
[…]
