### 2019-08-06 bd63c5f
Update simple_test.py
