### 2019-08-06 #1524 @alexanderimanicowenrivers
Dear @desilinguist,

Apologies, below are the final amendments. We are just working out how to add the badge from shields.io as we speak.

1. See https://github.com/for-ai/rl/pull/18
2. See https://github.com/for-ai/rl/pull/17
3. See https://github.com/for-ai/rl/pull/19

Let me know what you think.

Bests,
Alex
