### 2019-08-06 #1524 @desilinguist
@alexanderimanicowenrivers I think it would be better to review the changes all together after they are merged into the `master` branch.

In any case, I looked at the first PR for `setup.sh` and it looks like it hasn't really been tested with `macports`. I get the following error when running the command `port install qt open-mpi pkg-config ffmpeg`:

> Error: Port qt not found
