### 2019-08-08 ee91562
Update setup.sh (#27)
