### 2019-08-10 #1524 @alexanderimanicowenrivers
The major changes have now been merged, as well as the bug for macports has now been fixed in this PR - https://github.com/for-ai/rl/pull/27

Thanks @desilinguist

Alex
