### 2019-08-11 a02a292
remove-test-warnings (#19)
* Update train.py

* Update basic_test.py (#20)

* Update ddpg_test.py (#21)

* Update noisy_network_test.py (#22)

* Update noisy_network_test.py

* Update noisy_network_test.py

* Update noisy_network_test.py

* Alexanderimanicowenrivers patch 1 (#23)

* Update noisy_network_test.py

* Update noisy_network_test.py

* Update ppo_test.py

* Update simple_test.py

* Alexanderimanicowenrivers patch 1 (#24)

* Update noisy_network_test.py

* Update noisy_network_test.py

* Update prioritized_test.py
