### 2019-08-12 #1524 @desilinguist
Thanks @alexanderimanicowenrivers! It looks much, much better now. FYI, the echo command in the [PR](https://github.com/for-ai/rl/pull/27/files) doesn't match the actual command that's run so you should probably fix that :)

@mbobra please consider this approved from my side!
