### 2019-08-13 2048a5a
Update README.md
Adding JOSS badge
