+++ README.md
@@ -1,4 +1,4 @@
-## [FOR.ai](https://for.ai) Reinforcement Learning Codebase [![Build Status](https://travis-ci.org/for-ai/rl.svg?branch=master)](https://travis-ci.org/for-ai/rl)[![](https://img.shields.io/badge/-documentation-blue)](https://rl-codebase.readthedocs.io/en/latest/) 
+## [FOR.ai](https://for.ai) Reinforcement Learning Codebase [![status](http://joss.theoj.org/papers/a65da0f74f34be097b1c1189ae6abdc6/status.svg)](http://joss.theoj.org/papers/a65da0f74f34be097b1c1189ae6abdc6) [![Build Status](https://travis-ci.org/for-ai/rl.svg?branch=master)](https://travis-ci.org/for-ai/rl)[![](https://img.shields.io/badge/-documentation-blue)](https://rl-codebase.readthedocs.io/en/latest/) 
 Modular codebase for reinforcement learning models training, testing and visualization.
 
 **Contributors**: [Bryan M. Li](https://github.com/bryanlimy), [Alexander Cowen-Rivers](https://github.com/alexanderimanicowenrivers), [Piotr Kozakowski](https://github.com/koz4k), [David Tao](https://github.com/taodav), [Siddhartha Rao Kamalakara](https://github.com/srk97), [Nitarshan Rajkumar](https://github.com/nitarshan), [Hariharan Sezhiyan](https://github.com/hsezhiyan), [Sicong Huang](https://www.cs.toronto.edu/~huang/), [Aidan N. Gomez](https://github.com/aidangomez)
