### 2019-08-19 #1524 @alexanderimanicowenrivers
As you can see with our latest changes, the testing is also much clearer. I have managed to remove a majority of the tensorflow depreciated warnings. https://travis-ci.org/for-ai/rl
