### 2019-08-21 #1524 @mbobra
@alexanderimanicowenrivers Per the [JOSS editorial guidelines](https://joss.readthedocs.io/en/latest/editing.html#continued-attention-to-assigned-submissions), I need to give @paragkulkarni11 7-10 days to respond to Whedon's prompt. Sorry for the delay.
