### 2019-08-27 #1524 @alexanderimanicowenrivers
I have created a realise now :) - thanks https://github.com/for-ai/rl/releases
