### 2019-08-29 #1524 @mbobra
Great, thanks! Can you please [archive this release in Zenodo](https://guides.github.com/activities/citable-code/) and put the Zenodo badge in your README.md file?
