### 2019-09-01 #1524 @bryanlimy
Hi @mbobra I have uploaded our release to Zenodo and added the Zenodo badge on our README https://github.com/for-ai/rl/blob/master/README.md
