### 2019-09-10 #1524 @mbobra
@bryanlimy @alexanderimanicowenrivers It looks like the references section is not compiling properly. Can you please update the paper.bib file to add the URLs in the URL metadata field (and the DOIs in the DOI metadata field, if relevant)? You can take a look at an example [paper.bib](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) file here or the bib files of recently accepted papers [here](https://joss.theoj.org/papers/published).
