### 2019-09-13 #1524 @hsezhiyan
Hello @mbobra,

Thanks for your suggestion. We believe we fixed the issue in:

- [for-ai/rl#29](https://github.com/for-ai/rl/pull/29)

Can you please confirm it compiles properly now?
