### 2019-09-14 #1524 @bryanlimy
hi @mbobra I have just updated the DOI submission, please check if the paper.bib file is working properly. Thanks.
