### 2019-09-18 #1524 @mbobra
@bryanlimy Ah, I figured out why the references section is messed up. Can you please remove the tick marks? E.g. change
```
`@liang:2017`
```

to `@liang:2017` to produce Liang et al. (2017) or `[@Liang:2017]` to produce (Liang et al., 2017). I see that our documentation on this is a bit confusing -- sorry about that. Sorry that this last bit has taken a while, we're almost there 🌴
