### 2019-09-18 #1524 @hsezhiyan
Hi @mbobra ,

I've updated the `paper.md` in [PR#30](https://github.com/for-ai/rl/pull/30). Can you check if this fixes the problem?
