### 2019-09-18 #1524 @mbobra
@hsezhiyan The paper.bib file still does not include DOIs. What I mean is that there should be a value in the DOI metadata field. For example, one of the references in paper.md looks like this:

```
@article{caspi:2017,
  title={Reinforcement Learning Coach.(Dec. 2017)},
  author={Caspi, Itai and Leibovich, Gal and Novik, Gal},
  url={https://doi.org/10.5281/zenodo},
  volume={1134899},
  year={2017}
}
```
However, it should look like this:

```
@article{caspi:2017,
  title={Reinforcement Learning Coach.(Dec. 2017)},
  author={Caspi, Itai and Leibovich, Gal and Novik, Gal},
  doi={https://doi.org/10.5281/zenodo},
  volume={1134899},
  year={2017}
}
```
A DOI is different than a URL; a URL can point anywhere, but a DOI links specifically to https://doi.org. Some of the work you cited does have a DOI, but there is no DOI tag in your citations. Some (like the arXiv pre-prints) do not have DOIs, so you do not need to include a DOI tag. Please include DOIs for references that do have them.
