### 2019-09-18 #1524 @hsezhiyan
Hi @mbobra ,

Thanks for letting me know. This has been fixed in [PR#31](https://github.com/for-ai/rl/pull/31)
