### 2019-09-20 #1524 @mbobra
@hsezhiyan Can you please update the reference in paper.bib file as follows?

```
@article{caspi:2017,
  title={Reinforcement Learning Coach.(Dec. 2017)},
  author={Caspi, Itai and Leibovich, Gal and Novik, Gal},
  doi={10.5281/zenodo.1134899},
  url={https://doi.org/10.5281/zenodo.1134899},
  volume={1134899},
  year={2017}
}

```
