+++ paper.bib
@@ -29,7 +29,8 @@
 @article{caspi:2017,
   title={Reinforcement Learning Coach.(Dec. 2017)},
   author={Caspi, Itai and Leibovich, Gal and Novik, Gal},
-  doi={https://zenodo.org/record/1134899},
+  doi={10.5281/zenodo.1134899},
+  url={https://doi.org/10.5281/zenodo.1134899},
   volume={1134899},
   year={2017}
 }
