### 2019-09-30 #1524 @mbobra
@openjournals/joss-eics This paper is ready for acceptance! Nice work @hsezhiyan @alexanderimanicowenrivers @bryanlimy  🎉
