### 2019-10-01 #1524 @kyleniemeyer
Hi @alexanderimanicowenrivers, some fixes needed in the paper before we publish:

- [ ]  in the abstract, you have some undefined acronyms
- [ ]  in related work, you need some spaces between words and in-text citations (actually, I see this in a few places throughout the paper)
- [ ]  The in-text citation associated with "stable baselines" is not showing up correctly
- [ ]  on page 3, the citation following "Proximal Policy Optimization" is not showing up
- [ ]  the code/docs snippet related to train.py on page 3 has text that is cut off, so you should probably move those to new lines
- [ ] Similarly, the code snippet right above "Conclusions" has some text cut off
- [ ] The Abadi 2016 reference has some mistakes in the title (`{ }` that are showing up in the text)
- [ ] I don't think the author field is correct for the Keras-rl reference
- [ ] TF-agents reference is missing a year, and needs some other information—I can't tell what kind of reference this is. If software, it needs at minimum a URL
- [ ] The Pytorch reference is also missing some information (e.g., URL) and doesn't match the other software citations
- [ ] Silver 2014 reference: I assume this is a conference ref—if so, it is missing information.
- [ ] The Sutton 2000 reference should be:
```
@incollection{Sutton2000,
title = {Policy Gradient Methods for Reinforcement Learning with Function Approximation},
author = {Sutton, Richard S and David A. McAllester and Satinder P. Singh and Mansour, Yishay},
booktitle = {{Advances in Neural Information Processing Systems 12}},
editor = {S. A. Solla and T. K. Leen and K. M\"{u}ller},
pages = {1057--1063},
publisher = {MIT Press},
}
```
