### 2019-10-04 59f07f8
Fixing issues
