+++ paper.md
@@ -35,16 +35,16 @@ output:
 
 # Abstract 
 
-Vast reinforcement learning (RL) research groups, such as DeepMind and OpenAI, have their internal (private) reinforcement learning codebases, which enable quick prototyping and comparing of ideas to many SOTA methods. We argue the five fundamental properties of a sophisticated research codebase are; modularity, reproducibility, many RL algorithms pre-implemented, speed and ease of running on different hardware/ integration with visualization packages. 
+Vast reinforcement learning (RL) research groups, such as DeepMind and OpenAI, have their internal (private) reinforcement learning codebases, which enable quick prototyping and comparing of ideas to many state-of-the-art (SOTA) methods. We argue the five fundamental properties of a sophisticated research codebase are: modularity, reproducibility, many RL algorithms pre-implemented, speed and ease of running on different hardware/ integration with visualization packages. 
 Currently, there does not exist any RL codebase, to the author's knowledge, which contains all the five properties, particularly with TensorBoard logging and abstracting away cloud hardware such as TPU's from the user. The codebase aims to help distil the best research practices into the community as well as ease the entry access and accelerate the pace of the field. More detailed documentation can be found [here](https://rl-codebase.readthedocs.io/en/latest/).
 
 
 # Related Work
 
-There are currently various implementations available for reinforcement learning codebase like OpenAI baselines[@dhariwal:2017],  Stable baselines [@hill2019], Tensorforce[@schaarschmidt:2017], Ray rllib[@liang:2017], Intel Coach[@caspi:2017], Keras-RL [@kerasrl:2019], Dopamine baselines [@castro:2018] and TF-Agents [@guadarramatf]. Ray rllib [@liang:2017] is amongst the strongest of existing RL frameworks, supporting; distributed operations, TensorFlow[@abadi:2016], PyTorch[@paszke:2017] and multi-agent reinforcement learning (MARL). Unlike Ray rllib, we choose to focus on Tensorflow support, allowing us to integrate specific framework visualisation and experiment tracking into our codebase. On top of this, we are developing a Kuberenetes script for MacOS and Linux users to connect to any cloud computing platform, such as Google TPU’s, Amazon AWS etc. Most other frameworks are plagued with problems like usability issues (difficult to get started and increment over), very little modularity in code (no/ little hierarchy and code reuse), no asynchronous training support, weak support for TensorBoard logging and so on. All these problems are solved by our project, which is a generic codebase built for reinforcement learning (RL) research in Tensorflow[@schaarschmidt:2017], with favoured RL agents pre-implemented as well as integration with OpenAI Gym[@brockman:2016] environment focusing on quick prototyping and visualisation.
[…]
