### 2019-10-04 #1524 @hsezhiyan
Hi Kyle,

Thanks for you update! I apologize for the delayed response.

I've addressed your comments in [PR#32](https://github.com/for-ai/rl/pull/32)

I've fixed all issues except the Silver 2014 reference. I checked other papers on arxiv, and they cite they cite it the exact same way. Let me know if you have other concerns.
