### 2019-10-04 3ea5df9
Fixing references in paper.bib
