+++ paper.bib
@@ -73,15 +73,7 @@
   year={2017}
 }
 
-@inproceedings{sutton:2000,
-  title={Policy gradient methods for reinforcement learning with function approximation},
-  author={Sutton, Richard S and McAllester, David A and Singh, Satinder P and Mansour, Yishay},
-  booktitle={Advances in neural information processing systems},
-  pages={1057--1063},
-  year={2000}
-}
-
-@incollection{Sutton2000,
+@incollection{sutton2000,
   title = {Policy Gradient Methods for Reinforcement Learning with Function Approximation},
   author = {Sutton, Richard S and David A. McAllester and Satinder P. Singh and Mansour, Yishay},
   booktitle = {{Advances in Neural Information Processing Systems 12}},
@@ -92,10 +84,19 @@
 }
 
 @inproceedings{silver2014deterministic,
-  title={Deterministic policy gradient algorithms},
-  author={Silver, David and Lever, Guy and Heess, Nicolas and Degris, Thomas and Wierstra, Daan and Riedmiller, Martin},
-  booktitle={ICML},
-  year={2014}
+  title = 	 {Deterministic Policy Gradient Algorithms},
+  author = 	 {David Silver and Guy Lever and Nicolas Heess and Thomas Degris and Daan Wierstra and Martin Riedmiller},
+  booktitle = 	 {{Proceedings of the 31st International Conference on Machine Learning}},
+  pages = 	 {387--395},
+  year = 	 {2014},
+  editor = 	 {Eric P. Xing and Tony Jebara},
+  volume = 	 {32},
+  number =       {1},
+  series = 	 {Proceedings of Machine Learning Research},
+  address = 	 {Bejing, China},
+  month = 	 {22--24 Jun},
+  publisher = 	 {{PMLR}},
+  url = 	 {http://proceedings.mlr.press/v32/silver14.html},
 }
 
 @article{schulman:2017,
