### 2019-10-04 #1524 @kyleniemeyer
Hi @hsezhiyan, thanks for making those edits. I just submitted a PR that fixes two final things in the references: https://github.com/for-ai/rl/pull/33

- the Sutton 2000 reference was still missing some information; it looks like you added the BiBTeX citation I gave above, but your paper was referencing the original (incorrect) entry.
- For the Silver 2014 reference, it looks like the other papers on arXiv are not citing in correctly (and as a journal, we need to be a bit more rigorous with references); I found the paper at http://proceedings.mlr.press/v32/silver14.html, where they give the correct BibTeX citation info.
