### 2019-10-04 #1524 @hsezhiyan
@kyleniemeyer thanks for the review and the pull request! I have merged your PR.

Are we good to go?
